// src/common/constants/Enums/MessageType.ts
export enum MessageType {
    INCOMING_MESSAGE = 'message',
    AVAILABLE_ROOMS = 'availableRooms',
    CREATE_ROOM = 'createRoom',
    JOIN_ROOM = 'joinRoom',
    LEAVE_ROOM = 'leaveRoom',
    PLAYERS_IN_ROOM = 'playersInRoom',
    SET_USERNAME = 'setUserName',
    VALIDATE_USERNAME = "validateUserName",
    PLAYER_UPDATE = 'playerUpdate',
    GAME_START = 'gameStart',
    INIT_GAME = 'initGame',
    PLAYER_MISSION_CHANGE = 'missionChanged',
    GET_GAME_DETAILS = 'getGameDetails',
    START_VOTING = 'startVoting',
    VOTE_FOR_TEAM = 'voteForTeam',
    VOTE_MISSION = 'voteMission',
    RESET_MISSION = 'resetMission',
    MISSION_VOTES_RESULT = 'missionVotesResult',
    END_GAME = 'endGame',
    WAITING_ROOM_PLAYER_UPDATE = 'waitingRoomPlayerUpdate',
    WAITING_ROOM_LEAVE_ROOM = 'waitingRoomLeaveRoom'
}